requirements.txt
streamlit
pandas
sqlalchemy
python-dotenv
langchain
langchain-groq
groq
plotly